requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
transformers>=4.20.0
torch>=1.12.0
sentence-transformers>=2.2.0

# NVIDIA RAPIDS for GPU acceleration
cudf>=22.0.0
cuml>=22.0.0
cugraph>=22.0.0

# Graph processing
networkx>=2.8

# Document processing
python-docx>=0.8.11
jsonlines>=3.0.0

# LLM integration
ollama>=0.1.0
requests>=2.28.0

# Clustering and embeddings
umap-learn>=0.5.3
hdbscan>=0.8.28

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
loguru>=0.6.0

# Evaluation
datasets>=2.0.0
metrics>=0.3.0